Fixes incorrect value passed

+++ src/InterarrivalStats.cpp
@@ -42,7 +42,7 @@ void InterarrivalStats::updatePacketStats(struct timeval* when,
   if (update_iat && lastTime.tv_sec) {
     float deltaMS = Utils::msTimevalDiff(when, &lastTime);
 
-    if (deltaMS > 0) ndpi_data_add_value(&delta_ms, (u_int32_t)deltaMS);
+    if (deltaMS > 0) ndpi_data_add_value(&delta_ms, (u_int64_t)deltaMS);
   }
 
   memcpy(&lastTime, when, sizeof(struct timeval));
